afo_core/scripts/rag/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-qdrant>=0.1.0
langgraph>=0.0.20
qdrant-client>=1.7.0
python-frontmatter>=1.1.0
watchdog>=3.0.0
openai>=1.12.0

